+++ projects/services-layers/src/lib/types/Layers.ts
@@ -32,19 +32,36 @@ export interface ILayerStyleSet extends IAnyObject {
   content?: ILayerContent;
 }
 
+export interface IPopupParams {
+  layerId: string;
+  layerName: string;
+  mapEvent: any;
+  layer: any;
+  feature?: any;
+  color?: Uint8ClampedArray | Uint8Array | Float32Array | DataView | string;
+  /** properties of the feature/layer or asyncData */
+  properties?: IAnyObject
+}
+
 export interface popup {
   /** limit layer or feature properties: only those properties of a layer/feature, that are listed in this array, are being passed through to a popup-render-function */
   filterkeys?: Array<string>;
   /** To overwrite the keys (and only the keys) of the layer/feature properties. Object has the form {"oldKey": "newKey"} */
   properties?: IAnyObject;
-  /** function to create html string - popupobj: nativeLayer */
-  popupFunction?: (popupobj: IAnyObject) => string;
-  /** async function where you can paste a html string to the callback - popupobj: nativeLayer */
-  asyncPopup?: (popupobj: any, cb: (html: any) => void) => void;
+  /** 
+   * function to create the popup content.
+   * Return an HTML string or an object from which an HTML string is generated.
+   */
+  popupFunction?: (popupParams: IPopupParams) => string | IAnyObject;
+  /** 
+   * async function to create the popup content.
+   * Pass an HTML string, or an object from which an HTML string is generated, to the callback..
+   */
+  asyncPopup?: (popupParams: IPopupParams, cb: (content: string | IAnyObject) => void) => void;
   /** create popup using angular component */
   dynamicPopup?: {
     component: Type<any>;
-    getAttributes?: (args: any) => object;
+    getAttributes?: (args: IPopupParams) => object;
   };
   /** default event is click - use move for a popup on hover */
   event?: 'move' | 'click';
@@ -54,6 +71,8 @@ export interface popup {
   options?: IAnyObject;
   /** If the layer should be filtered out and the popup beneath should be shown e.g. text overlays use filterLayer: true */
   filterLayer?: boolean;
+  /** Use this to not add a popup to the map and instead publish events to view data in an external UI. */
+  asObservable?: boolean;
 }
 
 
@@ -127,7 +146,7 @@ export type TGeoExtent = [number, number, number, number] | [number, number, num
 export type CrossOriginType = 'anonymous' | 'use-credentials';
 
 
-export interface ILayerExpanded{
+export interface ILayerExpanded {
   /** tab: settings | legend | description */
   tab: string;
   /** optional to not expand the tab - for overriding defaults */
@@ -304,7 +323,7 @@ export class Layer implements ILayerOptions {
 
   styles?: ILayerStyleSet[];
   crossOrigin?: CrossOriginType;
-  expanded: boolean | ILayerExpanded  = false;
+  expanded: boolean | ILayerExpanded = false;
   cssClass?: string;
 
   constructor(options: ILayerOptions) {